+++ R/class-forecast-quantile.R
@@ -175,27 +175,57 @@ get_metrics.forecast_quantile <- function(x, select = NULL, exclude = NULL, ...)
 }
 
 
-#' @rdname get_pit
+#' @rdname get_pit_histogram
 #' @importFrom stats na.omit
 #' @importFrom data.table `:=` as.data.table
 #' @export
-get_pit.forecast_quantile <- function(forecast, by, ...) {
+get_pit_histogram.forecast_quantile <- function(forecast, num_bins = NULL,
+                                                breaks = NULL, by, ...) {
+  assert_number(num_bins, lower = 1, null.ok = TRUE)
+  assert_numeric(breaks, lower = 0, upper = 1, null.ok = TRUE)
   forecast <- clean_forecast(forecast, copy = TRUE, na.omit = TRUE)
   forecast <- as.data.table(forecast)
+  present_quantiles <- unique(c(0, forecast$quantile_level, 1))
+  present_quantiles <- round(present_quantiles, 10)
 
+  if (!is.null(breaks)) {
+    quantiles <- unique(c(0, breaks, 1))
+  } else if (is.null(num_bins) || num_bins == "auto") {
+    quantiles <- present_quantiles
+  } else {
+    quantiles <- seq(0, 1, 1 / num_bins)
+  }
+  ## avoid rounding errors
+  quantiles <- round(quantiles, 10)
+  diffs <- round(diff(quantiles), 10)
+
+  if (length(setdiff(quantiles, present_quantiles)) > 0) {
+    cli::cli_warn(
+      "Some requested quantiles are missing in the forecast. ",
+      "The PIT histogram will be based on the quantiles present in the forecast."
+    )
+  }
+
+  forecast <- forecast[quantile_level %in% quantiles]
   forecast[, quantile_coverage := (observed <= predicted)]
+
   quantile_coverage <-
     forecast[, .(quantile_coverage = mean(quantile_coverage)),
              by = c(unique(c(by, "quantile_level")))]
-  quantile_coverage <- quantile_coverage[
+
+  bins <- sprintf("[%s,%s)", quantiles[-length(quantiles)], quantiles[-1])
+  mids <- (quantiles[-length(quantiles)] + quantiles[-1]) / 2
+
+  pit_histogram <- quantile_coverage[
     order(quantile_level),
     .(
-      quantile_level = c(quantile_level, 1),
-      pit_value = diff(c(0, quantile_coverage, 1))
+      density = diff(c(0, quantile_coverage, 1)) / diffs,
+      bin = bins,
+      mid = mids
     ),
     by = c(get_forecast_unit(quantile_coverage))
   ]
-  return(quantile_coverage[])
+  return(pit_histogram[])
 }